Use hashed recovery codes instead
Yes, this will break tests. Yes, this may be slow. Yes, I may need to opt for pyca.bcrypt's KDF function. We will see. But in the name of science, break things (tests) to keep the codes safe.

+++ weasyl/two_factor_auth.py
@@ -6,6 +6,7 @@ from __future__ import absolute_import, unicode_literals
 import re
 import urllib
 
+import bcrypt
 import pyotp
 from qrcodegen import QrCode
 
@@ -114,7 +115,7 @@ def activate(userid, tfa_secret, tfa_response):
     """
     totp = pyotp.TOTP(tfa_secret)
     # If the provided `tfa_response` matches the TOTP value, write the 2FA secret into `login`, activating 2FA for `userid`
-    if totp.verify(tfa_response):
+    if totp.verify(tfa_response, valid_window=1):
         d.engine.execute("""
             UPDATE login
             SET twofa_secret = %(tfa_secret)s
@@ -215,6 +216,9 @@ def store_recovery_codes(userid, recovery_codes):
         if len(code) != LENGTH_RECOVERY_CODE:
             return False
 
+    # Store the recovery codes securely by hashing them with bcrypt (as login.passhash())
+    hashed_codes = [login.passhash(code) for code in codes]
+
     # If above checks have passed, clear current recovery codes for `userid` and store new ones
     d.engine.execute("""
         BEGIN;
@@ -226,7 +230,7 @@ def store_recovery_codes(userid, recovery_codes):
         SELECT %(userid)s, unnest(%(tfa_recovery_codes)s);
 
         COMMIT;
-    """, userid=userid, tfa_recovery_codes=list(codes))
+    """, userid=userid, tfa_recovery_codes=list(hashed_codes))
 
     # Verify if the atomic transaction completed; if `code` (one of the new recovery codes) is
     #   valid at this point, the new codes were added
@@ -257,25 +261,27 @@ def is_recovery_code_valid(userid, tfa_code, consume_recovery_code=True):
     # Recovery codes must be LENGTH_RECOVERY_CODE characters; fast-fail if this is not the case
     if len(tfa_code) != LENGTH_RECOVERY_CODE:
         return False
-    # Check to see if the provided code--converting to upper-case first--is valid and consume if so
-    if consume_recovery_code:
-        tfa_rc = d.engine.scalar("""
-            DELETE FROM twofa_recovery_codes
-            WHERE userid = %(userid)s AND recovery_code = %(recovery_code)s
-            RETURNING recovery_code
-        """, userid=userid, recovery_code=tfa_code.upper())
-    else:
-        # We only want to see if the code is valid at the moment.
-        tfa_rc = d.engine.scalar("""
-            SELECT recovery_code
-            FROM twofa_recovery_codes
-            WHERE userid = %(userid)s AND recovery_code = %(recovery_code)s
-        """, userid=userid, recovery_code=tfa_code.upper())
-    # Return True if the recovery code was valid, False otherwise
-    if tfa_rc:
-        return True
-    else:
-        return False
+    
+    # First extract the bcrypt hashes.
+    recovery_codes = d.engine.execute("""
+        SELECT recovery_code
+        FROM twofa_recovery_codes
+        WHERE userid = %(userid)s AND recovery_code = %(recovery_code)s
+    """, userid=userid, recovery_code=tfa_code.upper()).fetchall()
+    
+    # Then attempt to hash the input code versus the stored code(s).
+    for recovery_code_hash in recovery_codes:
+        if bcrypt.checkpw(tfa_code, recovery_code_hash):
+            # We have a match! If we are deleting the code, do it now.
+            if consume_recovery_code:
+                d.engine.execute("""
+                    DELETE FROM twofa_recovery_codes
+                    WHERE userid = %(userid)s AND recovery_code = %(recovery_code)s
+                """, userid=userid, recovery_code=recovery_code_hash)
+            # After deletion--if applicable--return that we succeeded.
+            return True
+    # If we get here, ``tfa_code`` did not match any stored codes. Return that we failed.
+    return False
 
 
 def is_2fa_enabled(userid):
